docs: Improve troubleshooting documentation
Due to the temporarily doubled ESP space usage, it is now easier to run
into the out of space issue (once). Document how to proceed in this case
without having to delete any generations.

Furthermore, recovery in case of ESP corruption is now slightly more
involved, because not all files are rewritten all the time. Adjust the
documentation accordingly.