tool(shared): introduce a build esp paths implementation
Usually, we build towards a certain ESP in our model.
Sometimes, we don't know in advance our ESP or don't care about this (building a stub that can be streamed, netbooted,
etc.).

This is a trivial implementation of a set of "build tree" ESP paths, e.g. no GC and only a root path where
everything is in together.