uefi(thin-stub): wip! wip! wip! transfer control via PXE APIs
In situations where we are not booted by a local filesystem, the SFS won't be available.

But we can try to look if we have been booted by a PXE API and get the protocol
to look for remote files that can enable us to continue the boot process while not breaking the
Secure Boot guarantees.