feat: add boot-counting support to lanzaboote
See https://systemd.io/AUTOMATIC_BOOT_ASSESSMENT/

Boot counting is controlled with a single option, `boot.lanzaboote.bootCounting.initialTries`,
if this option is set to a non-zero value, new boot entries will be created with
a counter added, set to the value specified in that option.
If the option is set to zero, which is also the default,
then no boot-counting counters will be added.
Already existing entries will not be modified,
so boot counting only applies to new boot entries.